mylist

+++ src/misc/qwen-axdr.c
@@ -561,7 +561,7 @@ bool axdr_encode_bitstring_fixed(AxdrBuffer *buf, const uint8_t *bits, int64_t b
  * -------------------------------------------------------------------------------------------
  * @return 成功 - 返回 true; 失败 - 返回 false
  ********************************************************************************************/
-bool axdr_decode_bitstring_fixed(AxdrBuffer *buf, uint8_t *bits, int64_t *bit_count, int64_t bitLengh)
+bool axdr_decode_bitstring_fixed(AxdrBuffer *buf, uint8_t *bits, uint64_t *bit_count, int64_t bitLengh)
 {
     if (buf->error || !bits || !bit_count || bitLengh < 0 || buf->pos + 1 > buf->size)
     {
@@ -652,7 +652,7 @@ bool axdr_encode_bitstring_var(AxdrBuffer *buf, const uint8_t *bits, int64_t bit
  * -------------------------------------------------------------------------------------------
  * @return 成功 - 返回 true; 失败 - 返回 false
  ********************************************************************************************/
-bool axdr_decode_bitstring_var(AxdrBuffer *buf, const uint8_t *bits, uint64_t bits_buffer_size, int64_t *bit_count)
+bool axdr_decode_bitstring_var(AxdrBuffer *buf, const uint8_t *bits, uint64_t bits_buffer_size, uint64_t *bit_count)
 {
     if (buf->error || !bits || bits_buffer_size <= 0 || !bit_count)
     {
@@ -673,7 +673,7 @@ bool axdr_decode_bitstring_var(AxdrBuffer *buf, const uint8_t *bits, uint64_t bi
     }
 
     *bit_count = len;
-    bool result =  axdr_decode_bitstring_fixed(buf, (uint8_t*) bits, bit_count, len);
+    bool result = axdr_decode_bitstring_fixed(buf, (uint8_t*) bits, bit_count, len);
     if (!result || len != *bit_count || (len + 7) / 8 > bits_buffer_size)
     {
         *bit_count = 0;
@@ -1149,7 +1149,7 @@ void test_struct_decoder(AxdrBuffer *buf, void *data)
         return;
 
     // 5. BitString
-    int bit_count;
+    uint64_t bit_count;
     if (!axdr_decode_bitstring_fixed(buf, ts->bitstring_val, &bit_count, 16))
         return;
 
@@ -1244,31 +1244,33 @@ void test_bitstring_codec(void)
         bool should_pass;    // Whether test should pass
         const char *desc;    // Test description
     } tests[] = {
-        // 测试1：单个字节，8位全部使用 (10000001)
-        {{0x81, 0x00, 0x00}, 8, true, "Single byte (10000001)"},
-        
-        // 测试2：单个位在最左侧 (10000000)
-        {{0x80, 0x00, 0x00}, 1, true, "Single bit at MSB (1)"},
-        
-        // 测试3：12位跨两个字节 (101010101010)
-        {{0xAA, 0xA0, 0x00}, 12, true, "12 bits across two bytes (101010101010)"},
-        
-        // 测试4：16位完整两个字节 (1010010110100101)
-        {{0xA5, 0xA5, 0x00}, 16, true, "16 bits full two bytes (1010010110100101)"},
-        
-        // 测试5：24位三个字节 (111111110000000011111111)
-        {{0xFF, 0x00, 0xFF}, 24, true, "24 bits three bytes (111111110000000011111111)"},
-        
-        // 测试6：无效位数
-        {{0x00, 0x00, 0x00}, -1, false, "Invalid bit count"}
+                  // 测试1：单个字节，8位全部使用 (10000001)
+            { { 0x81, 0x00, 0x00 }, 8, true, "Single byte (10000001)" },
+
+            // 测试2：单个位在最左侧 (10000000)
+            { { 0x80, 0x00, 0x00 }, 1, true, "Single bit at MSB (1)" },
+
+            // 测试3：12位跨两个字节 (101010101010)
+            { { 0xAA, 0xA0, 0x00 }, 12, true, "12 bits across two bytes (101010101010)" },
+
+            // 测试4：16位完整两个字节 (1010010110100101)
+            { { 0xA5, 0xA5, 0x00 }, 16, true, "16 bits full two bytes (1010010110100101)" },
+
+            // 测试5：24位三个字节 (111111110000000011111111)
+            { { 0xFF, 0x00, 0xFF }, 24, true, "24 bits three bytes (111111110000000011111111)" },
+
+            // 测试6：无效位数
+            { { 0x00, 0x00, 0x00 }, -1, false, "Invalid bit count" }
     };
 
-    for (size_t i = 0; i < sizeof(tests)/sizeof(tests[0]); i++) {
+    for (size_t i = 0; i < sizeof(tests) / sizeof(tests[0]); i++)
+    {
         printf("\nTest Case %zu: %s\n", i + 1, tests[i].desc);
-        
+
         // Create encoder buffer
         AxdrBuffer *enc_buf = axdr_buffer_new_encoder(32);
-        if (!enc_buf) {
+        if (!enc_buf)
+        {
             printf("Failed to create encoder buffer\n");
             continue;
         }
@@ -1276,76 +1278,93 @@ void test_bitstring_codec(void)
         // Encode
         bool enc_result = axdr_encode_bitstring_fixed(enc_buf, tests[i].bits, tests[i].bit_count);
         printf("Encoding %s\n", enc_result ? "succeeded" : "failed");
-        
-        if (enc_result != tests[i].should_pass) {
+
+        if (enc_result != tests[i].should_pass)
+        {
             printf("Unexpected encoding result!\n");
             axdr_buffer_free(enc_buf);
             continue;
         }
 
-        if (!enc_result) {
+        if (!enc_result)
+        {
             axdr_buffer_free(enc_buf);
             continue;
         }
 
         // Print encoded data
         printf("Encoded data (%zu bytes): ", enc_buf->pos);
-        for (size_t j = 0; j < enc_buf->pos; j++) {
+        for (size_t j = 0; j < enc_buf->pos; j++)
+        {
             printf("%02X ", enc_buf->data[j]);
         }
         printf("\n");
 
         // Print bit pattern for verification
-        if (tests[i].bit_count > 0) {
+        if (tests[i].bit_count > 0)
+        {
             printf("Bit pattern: ");
-            for (int j = 0; j < tests[i].bit_count; j++) {
+            for (int j = 0; j < tests[i].bit_count; j++)
+            {
                 int byte_index = j / 8;
                 int bit_index = 7 - (j % 8);  // Start from MSB
                 printf("%d", (tests[i].bits[byte_index] >> bit_index) & 0x01);
-                if ((j + 1) % 8 == 0) printf(" ");
+                if ((j + 1) % 8 == 0)
+                    printf(" ");
             }
             printf("\n");
         }
 
         // Create decoder buffer
         AxdrBuffer *dec_buf = axdr_buffer_new_decoder(enc_buf->data, enc_buf->pos);
-        if (!dec_buf) {
+        if (!dec_buf)
+        {
             printf("Failed to create decoder buffer\n");
             axdr_buffer_free(enc_buf);
             continue;
         }
 
         // Decode
-        uint8_t decoded_bits[3] = {0};
-        int decoded_bit_count = 0;
+        uint8_t decoded_bits[3] = { 0 };
+        uint64_t decoded_bit_count = 0;
         bool dec_result = axdr_decode_bitstring_fixed(dec_buf, decoded_bits, &decoded_bit_count, tests[i].bit_count);
         printf("Decoding %s\n", dec_result ? "succeeded" : "failed");
 
-        if (dec_result) {
+        if (dec_result)
+        {
             // Print decoded bit pattern
             printf("Decoded pattern: ");
-            for (int j = 0; j < decoded_bit_count; j++) {
+            for (int j = 0; j < decoded_bit_count; j++)
+            {
                 int byte_index = j / 8;
                 int bit_index = 7 - (j % 8);  // Start from MSB
                 printf("%d", (decoded_bits[byte_index] >> bit_index) & 0x01);
-                if ((j + 1) % 8 == 0) printf(" ");
+                if ((j + 1) % 8 == 0)
+                    printf(" ");
             }
             printf("\n");
 
             // Verify results
             bool match = true;
-            if (decoded_bit_count != tests[i].bit_count) {
+            if (decoded_bit_count != tests[i].bit_count)
+            {
                 match = false;
-            } else {
-                for (int j = 0; j < (decoded_bit_count + 7) / 8; j++) {
+            }
+            else
+            {
+                for (int j = 0; j < (decoded_bit_count + 7) / 8; j++)
+                {
                     uint8_t mask = 0xFF;
-                    if (j == (decoded_bit_count + 7) / 8 - 1) {
+                    if (j == (decoded_bit_count + 7) / 8 - 1)
+                    {
                         int remaining_bits = decoded_bit_count % 8;
-                        if (remaining_bits != 0) {
+                        if (remaining_bits != 0)
+                        {
                             mask = 0xFF << (8 - remaining_bits);
                         }
                     }
-                    if ((decoded_bits[j] & mask) != (tests[i].bits[j] & mask)) {
+                    if ((decoded_bits[j] & mask) != (tests[i].bits[j] & mask))
+                    {
                         match = false;
                         break;
                     }
@@ -1368,9 +1387,11 @@ void axdrtest(void)
 
     // 1. 准备测试数据
     TestStruct original_data = { .int16_val = -12345, .uint16_val = 54321, .bool_val = true, .enum_val = 2, .bitstring_val = { 0xAB, 0xCD },
-            .octetstring_val = { 0x11, 0x22, 0x33, 0x44 }, .octetstring_len = 4, .visiblestring_val = strdup("Hello, A-XDR!"), .time_val = time(NULL),
-            .seq_of_count = 3, .seq_of_vals = NULL
-                    };
+                                 .octetstring_val = { 0x11, 0x22, 0x33, 0x44 },
+                                 .octetstring_len = 4, .visiblestring_val = strdup("Hello, A-XDR!"), .time_val = time(NULL),
+                                 .seq_of_count = 3,
+                                 .seq_of_vals = NULL
+    };
     original_data.seq_of_vals = malloc(3 * sizeof(int32_t));
     if (!original_data.seq_of_vals)
     {
